Move GetMountType to dos_mscdex file
This was the only user of this function.
